IS-1966: Relocated nokkelinformasjon folder
